services/api/requirements.txt
fastapi[standard]
pydantic-settings
openai>=1.40.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.0
pandas>=2.2.2
duckduckgo-search>=6.1.6
httpx>=0.27.0
trafilatura>=1.9.0
beautifulsoup4>=4.12.3
pydantic>=2.7.0
sqlalchemy
psycopg[binary]
pgvector
uvicorn[standard]
